movies_admin/requirements.txt
django==3.2
flake8==4.0
python-dotenv==1.0.0
split_settings==1.0.0
django-debug-toolbar==3.4.0
psycopg2-binary==2.9
uwsgi==2.0.20
django-cors-headers==3.14.0